02-fastapi-backend/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-generativeai
python-dotenv